requirements.txt
absl-py==1.0.0
appdirs==1.4.4
astunparse==1.6.3
attrs==21.4.0
audioread==2.1.9
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cycler==0.11.0
decorator==5.1.1
flatbuffers==2.0
fonttools==4.30.0
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
htmlmin==0.1.12
idna==3.3
ImageHash==4.2.1
importlib-metadata==4.11.3
Jinja2==3.0.3
joblib==1.0.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
libclang==13.0.0
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
missingno==0.5.1
multimethod==1.7
networkx==2.7.1
numba==0.55.1
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandas-profiling==3.1.0
phik==0.12.1
Pillow==9.0.1
pooch==1.6.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
tangled-up-in-unicode==0.1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.9
visions==0.7.4
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0
